moving shared steps to separate files in Step02
